requirements.txt
python-dotenv==1.1.1
fastapi==0.116.1
redis==6.4.0
crewai==0.165.1
python-docx==1.2.0
langchain==0.3.27
langchain-openai==0.3.31
langchain-community==0.3.27
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
PyJWT==2.10.1
werkzeug==3.1.3
reportlab==4.4.3
apscheduler==3.11.0
azure-storage-blob==12.26.0
azure-cosmos==4.9.0
azure-functions==1.23.0
streamlit==1.48.1
pypdf==6.0.0
requests==2.32.5
beautifulsoup4
duckduckgo-search
